requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via typer
contourpy==1.1.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.12.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.43.1
    # via matplotlib
fsspec==2023.9.2
    # via
    #   huggingface-hub
    #   torch
geomloss==0.2.6
    # via opendataval (pyproject.toml)
huggingface-hub==0.18.0
    # via
    #   tokenizers
    #   transformers
idna==3.4
    # via requests
jinja2==3.1.2
    # via torch
joblib==1.3.2
    # via scikit-learn
keopscore==2.1.2
    # via pykeops
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.3
    # via jinja2
matplotlib==3.8.0
    # via opendataval (pyproject.toml)
mpmath==1.3.0
    # via sympy
multimethod==1.10
    # via pandera
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.1
    # via torch
numpy==1.25.2
    # via
    #   contourpy
    #   geomloss
    #   matplotlib
    #   opendataval (pyproject.toml)
    #   pandas
    #   pandera
    #   pykeops
    #   scikit-learn
    #   scipy
    #   torchvision
    #   transformers
packaging==23.2
    # via
    #   huggingface-hub
    #   matplotlib
    #   pandera
    #   transformers
pandas==1.5.3
    # via
    #   opendataval (pyproject.toml)
    #   pandera
pandera==0.15.2
    # via opendataval (pyproject.toml)
pillow==10.0.1
    # via
    #   matplotlib
    #   torchvision
pybind11==2.11.1
    # via pykeops
pydantic==1.10.13
    # via pandera
pykeops==2.1.2
    # via opendataval (pyproject.toml)
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
regex==2023.10.3
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   opendataval (pyproject.toml)
    #   torchvision
    #   transformers
safetensors==0.4.0
    # via transformers
scikit-learn==1.3.1
    # via opendataval (pyproject.toml)
scipy==1.11.3
    # via
    #   opendataval (pyproject.toml)
    #   scikit-learn
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
threadpoolctl==3.2.0
    # via scikit-learn
tokenizers==0.15.0
    # via transformers
# torch==2.1.0
    # via
    #   geomloss
    #   opendataval (pyproject.toml)
    #   torchvision
torcheval==0.0.7
    # via opendataval (pyproject.toml)
# torchvision==0.16.0
    # via opendataval (pyproject.toml)
tqdm==4.64.1
    # via
    #   huggingface-hub
    #   opendataval (pyproject.toml)
    #   transformers
transformers==4.35.2
    # via opendataval (pyproject.toml)
typeguard==4.1.5
    # via pandera
typer==0.9.0
    # via opendataval (pyproject.toml)
typing-extensions==4.8.0
    # via
    #   huggingface-hub
    #   pydantic
    #   torch
    #   torcheval
    #   typeguard
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via pandera
urllib3==2.0.6
    # via requests
wrapt==1.15.0
    # via pandera
